requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.3.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.1
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.1
Pillow==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
toolz==0.9.0
uritemplate==3.0.0
urllib3==1.24.1
